each module should have its own admin
the admin config in facebook/admin.py was used for only display used
modules in the admin. now, we have separated modules, so we can have
normal admins.
